Refactor: Reorganize classes into ecmwf and meteofrance submodules
- Create ecmwf and meteofrance submodules.
- Move model classes into the appropriate submodule.
- Update imports to reflect the new structure.
- Ensure backward compatibility by exposing models in the top-level __init__.py.
